0.1.0
added a config/connection.js to create a new instance of sequelize server that will uses JASWDB if exits else it will create an instance of the sequelize server that uses process env vars for name, user, pass. then model is exported.  Created schema. sql to drop the db if exists and CREATE DATABASE to create the same db. Created a models/user.js model, that defines the user model and has 2 built in methods one to return the hashed with bcrypt pass, and the other to return the usernaame, accepts data for id, user_name, email, password, uses hook for password beforeCreate to bycrpyt hash the password. created index.js to manage relationships between models and export them as a group. Added data to server js to use express, express session path and sequelizestore, created a sess object that creates a session and a cookie and stores that in the sequelize database, set the server to use the session with the argument of the sess object. const set exhbs to create helpers from utils/helpers.js set the app.engine to use hadnlebars. then finally the sequelize.sync syncs the models then uses a  .then and a callback to start the server listening on a port or the process.enviornment port