0.17.0
added a create user route that creates a user by properties username email password and then saves a session with the logged in param as true and the current user as the request bodys username. if all is well 200+ userdata  else  500 + errcode, added a loggin route that uses model user and find one method on model where email is equal to request body email  validates the user was found  then defines verify pass as the found user instance and calls the class method of the use model that uses bcrypt to check if the req.body password is equal to the saved pass. validates verify pass if all is well the session is saved with the loggedIn variable as true and then creates a current user variable that is equal to the class method userName() which returns the username. and response is 200+ message  if there was an errror 500+ errcode