0.21.0
Added in ability to see reviews relating to the relative post.  added in a get req that uses /park/:id and itterates over the park model to find a primary key that is equal to the req.params.id and includes reviews related. maps the reviews and renders, singlepark.handlebars with object data name, counter, image, content, reviews and loggedIn status. if all went well. res 200 + render,  if catch 500+ err code. updated seed/data to have a diffrent post for testing. updated homepage to make each park a link that directs to /parks/parkid , added in foreach statment to itterate over all reviews, added in data for select park