0.37.3
on main.handlebars I fixed the order of the stylesheets by putting the bootstrap link over the css sheet so  styling wont interfere. I also added a sign up button to the nav bar when the user is logged out. For the singlepark.handlebars I changed the order of the park reviews and the review form by putting the review form over the park reviews for a better user experience. for the main.handlebars page I fixed a lot of bootstrap errors and class names and I got rid of head and html tags which werent supposed to be there. On the temp.css I added styling to the homepage and main.handlebar pages. Lastly for the parkData.js I fixed some of the image links, I put it in alphabetical order by name, and I deleted some historic sites that were not supposed to be there.